implement S3 CORS, move validation into file
